+++ src/main/java/lotto/model/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.util.Collection;
 import java.util.Objects;
