+++ src/main/java/lotto/model/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.util.Arrays;
 import java.util.List;
