+++ src/main/java/lotto/model/Money.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.math.BigDecimal;
 import java.math.MathContext;
