+++ src/test/java/lotto/model/MoneyTest.java
@@ -1,9 +1,11 @@
-package lotto;
+package lotto.model;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+
+import lotto.model.Money;
 
 public class MoneyTest {
