+++ src/test/java/lotto/model/RankTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
